requirements.txt
langchain-openai
streamlit
python-dotenv
openai
langchain
langchain-community
chromadb
pypdf
tiktoken
